+++ CHANGES.md
@@ -110,6 +110,7 @@
 * Add new test for wanted whole first season (add show)
 * Change SickGear-NG version
 * Fix "Update shows on startup" (change update_shows_on_start = 0 in config.ini, restart SG, then update)
+* Fix Plex notifier, PY2 urllib2.urlopen has no `with` context manager and six.moves.urllib.request.urlopen does not provide
 
 
 [develop changelog]